requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
pip==22.3
protobuf==3.20.1
python-dotenv==0.21.0
requests==2.28.1
setuptools==65.3.0
soupsieve==2.3.2.post1
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
